pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "n4cluster-icp-finder"
version = "0.1.0"
description = "Restaurant ICP identification and scoring system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    "celery[redis]>=5.4.0",
    "playwright>=1.45.0",
    "scrapy>=2.11.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "pydantic-settings>=2.5.0",
    "scikit-learn>=1.5.0",
    "hdbscan>=0.8.38",
    "numpy>=1.26.0",
    "openai>=1.40.0",
    "anthropic>=0.34.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "ruff>=0.5.0",
]

[project.scripts]
icp-server = "src.main:run"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 120
